source/public_api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncmock==0.4.2
asyncpg==0.29.0
attrs==23.2.0
boto3==1.34.131
botocore==1.34.131
botocore-stubs==1.34.143
caio==0.9.17
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
docker==7.1.0
dump-env==1.5.0
email_validator==2.2.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fixtures==4.1.0
frozenlist==1.4.1
fs==2.4.16
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
loguru==0.7.2
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
minio==7.2.7
miniopy-async==1.20.1
mock==5.1.0
multidict==6.0.5
orjson==3.10.5
packaging==24.1
pbr==6.0.0
pillow==10.4.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-asyncio==0.23.8
pytest-dotenv==0.5.2
pytest-env==1.1.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
requests-toolbelt==1.0.0
responses==0.25.3
rich==13.7.1
s3transfer==0.10.2
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
SQLAlchemy==2.0.31
starlette==0.37.2
structlog==24.2.0
testcontainers==4.3.3
testcontainers-core==0.0.1rc1
testcontainers-postgres==0.0.1rc1
tomli==2.0.1
typer==0.12.3
types-awscrt==0.21.0
types-cryptography==3.3.23.2
types-s3transfer==0.10.1
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
